saas/requirements.txt
django>=5.1,<6
djangorestframework>=3.15,<4
django-allauth>=65.0,<66
psycopg2-binary>=2.9,<3
stripe>=10.0,<12
whitenoise>=6.7,<7
python-decouple>=3.8,<4
dj-database-url>=2.1,<3
gunicorn>=22.0,<23
